requirements.txt
nicegui==2.3.0
openai==1.51.2
Pillow==10.4.0
replicate==0.34.2
Requests==2.32.3
together==1.3.1
transformers==4.45.2
